+++ nf_core/pipelines/refgenie.py
@@ -181,13 +181,17 @@ def update_config(rgc):
         log.info("Could not determine path to 'refgenie_genomes.config' file.")
         return False
 
+    if refgenie_genomes_config_file is None:
+        log.info("Could not determine path to 'refgenie_genomes.config' file.")
+        return False
+
     # Save the updated genome config
     try:
         with open(refgenie_genomes_config_file, "w+") as fh:
             fh.write(refgenie_genomes)
         log.info(f"Updated nf-core genomes config: {refgenie_genomes_config_file}")
     except FileNotFoundError:
-        log.warning(f"Could not write to {refgenie_genomes_config_file}")
+        log.info(f"Could not write to {refgenie_genomes_config_file}")
         return False
 
     # Add include statement to NXF_HOME/config